+++ pozdnyakov.artemy/B3/iterator.hpp
@@ -5,26 +5,26 @@
 
 namespace pozdnyakov
 {
-  class FactorialIterator: public std::iterator< std::bidirectional_iterator_tag, size_t >
+  class FactorialIterator: public std::iterator< std::bidirectional_iterator_tag, unsigned long >
   {
   public:
     FactorialIterator& operator++();
     FactorialIterator operator++(int);
     FactorialIterator& operator--();
     FactorialIterator operator--(int);
-    size_t& operator*();
-    size_t* operator->();
+    unsigned long& operator*();
+    unsigned long* operator->();
     bool operator==(const FactorialIterator& it) const;
     bool operator!=(const FactorialIterator& it) const;
   private:
     friend class FactorialContainer;
-    static const size_t MAX_NUM = 10;
-    size_t n_;
-    size_t value_;
+    static const unsigned MAX_NUM = 10;
+    unsigned n_;
+    unsigned long value_;
     FactorialIterator() = default;
-    FactorialIterator(size_t n, size_t value);
+    FactorialIterator(unsigned n, unsigned long value);
   };
-  size_t getFactorial(size_t n);
+  unsigned long getFactorial(unsigned n);
 }
 
 #endif
